requirements.txt
torch
torchvision
numpy
scipy
pillow
numpy
pandas
matplotlib
scikit-learn
scikit-image
fire